@@ -1,10 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   /* config options here */
-  // reactCompiler: true,
+  reactCompiler: true,
   // serverExternalPackages: ["next-auth"],
 
-  reactStrictMode: true,
+  // reactStrictMode: true,
 };
 
-module.exports = nextConfig;
+export default nextConfig;
